32bit support

+++ src/arch/x86_64/time.rs
@@ -1,3 +1,3 @@
-pub fn get_time() -> u64 {
+pub fn get_time() -> usize {
     unsafe { core::arch::x86_64::_rdtsc() }
 }
