+++ src/protocol/input.rs
@@ -7,7 +7,7 @@ pub struct InputEvent {
     pub kind: u8,
     pub code: u16,
     pub value: i32,
-    pub timestamp: u64,
+    pub timestamp: usize,
 }
 
 pub const INPUT_OP_READ: u8 = 0x1;
